Refactored Stake object
